move makePromise.js out of src/kernel/ up into src/
